Updated the way items are stored in inventories to work with tools and polymorphism.

+++ SNESMimicTypeDealy/!TestGame/Items/Types/Tool.cpp
@@ -2,7 +2,7 @@
 
 namespace GAME_NAME::Items
 {
-	Tool::Tool(ITEM_TYPE type, float uses)
+	Tool::Tool(ITEM_TYPE type, int uses)
 		: InventoryItem(type), m_uses(uses)
 	{
